requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.4
deepface==0.0.96
Django==5.2
django-channels==0.7.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
easydict==1.13
exceptiongroup==1.3.1
filelock==3.20.2
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
inflection==0.5.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
kombu==5.6.2
lazy_loader==0.4
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.2.0
rpds-py==0.30.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.12
six==1.17.0
sounddevice==0.5.3
soupsieve==2.8.1
sqlparse==0.5.4
stringzilla==4.6.0
swagger-spec-validator==3.0.4
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==2.0.1
